Tests/LibWeb/Text/expected/ResizeObserver/observe-after-load-fires-initial-callback.txt

width=200 height=200
